Moves to new folder
