+++ lib/train/transports/helpers/azure/file_parser.rb
