+++ lib/train/transports/helpers/azure/subscription_id_file_parser.rb
