+++ lib/train/transports/helpers/azure/subscription_number_file_parser.rb
